requirements.txt
# ============================================
# SIPROSA MES - Dependencias Python
# ============================================

# Framework Django
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

# Base de datos
psycopg2-binary==2.9.10

# CORS y configuración
django-cors-headers==4.9.0
python-dotenv==1.1.1

# Servidor de producción
gunicorn==23.0.0

# Utilidades
PyJWT==2.10.1
sqlparse==0.5.3
typing-extensions==4.15.0
